forms finish

+++ app/forms/create_product.py
@@ -1,18 +1,18 @@
 from flask_wtf import FlaskForm
-from wtforms import IntegerField, StringField, FloatField, SelectField
+from wtforms import IntegerField, StringField, FloatField, SelectField, DecimalField
 from wtforms.validators import DataRequired()
 
 class CreateProductForm(FlaskForm):
     product_name = StringField('Name', validators=[DataRequired()])
     description = StringField('Description', validators=[DataRequired()])
-    price = IntegerField('Price', validators=[DataRequired()])
+    price = DecimalField('Price', validators=[DataRequired()])
     category = SelectField('Category', choices=['Cars', 'Clothing', 'Electronics', 'Home Goods', 'miscellaneous'], validators=[DataRequired()])
     stock = IntegerField('Stock', validators=[DataRequired()])
 
 class EditProductForm(FlaskForm):
     product_name = StringField('Name', validators=[DataRequired()])
     description = StringField('Description', validators=[DataRequired()])
-    price = IntegerField('Price', validators=[DataRequired()])
+    price = DecimalField('Price', validators=[DataRequired()])
     category = SelectField('Category', choices=['Cars', 'Clothing', 'Electronics', 'Home Goods', 'miscellaneous'], validators=[DataRequired()])
     stock = IntegerField('Stock', validators=[DataRequired()])
